exercise completed with minimal differences
I was able to replicate the desired outcome screenshot with the only difference in my css's is that I used flexbox to center my text and the solution used "align-text:center;". I learned that even though there's nothing wrong with using flexbox, if I want to horizontally center text or inline elements, "align-text" is a quicker way of getting the job done.

flexbox is still a good go-to for almost any situation