Exercise completed 09/02/2024 9:40pmPST
I was stumped because I did not know or maybe forgot how to use the flex property.

flex: ; allows for the element to grow to fill the available space. So the content div in my code needed to grow in order for the header and footer to be stuck in the desired position.