Exercise completed 9/3/2024 4:51pm
Took a little bit of brain power, but was able to complete the task with minimal struggle.
Instead of adding "width" property to each card I added flex basis. I should've used width property Instead
because now the element will always start as being displayed with 300px as the width